Attendance System Update
Added a feature where all students are automatically inserted into Student Attendance tables. This means that a student's attendance is "absent" by default. It prevents an officer from having to miss marking a student's attendance.

This feature is also vital for the future iteration of Community Service automation. It will enable the system to automatically count the number of absences of students. Future iteration will also include an individual viewing of attendance (for Student users)

The attendance view will be seen through visiting an event's details from the Events List page. The user must select a certain date to be able to view the list of students' attendance. By doing this kind of system design, we can reduce the redundancy of information inside the Events List table (i.e. duplicate data).

An attendance will always have at least one and a maximum of two types with one entry of each: AM and PM.

As observed from the previous events, SBO's attendance will always two signings regardless of what kind of event it is: Sign In and Sign Out. An event's attendance may also happen only once, hence, the optional addition of AM/PM types.